requirements.txt
rich==13.6.0
langchain==0.0.317
google-cloud-aiplatform==1.35.0
sentence-transformers==2.2.2